requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0

# Transformers (for BART embeddings)
transformers>=4.30.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Development (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
